Split `crates/fuzzing/tests/oom.rs` into multiple modules
It was getting big and unwieldy.
